pyproject.toml
[project]
name = "pet-simulation"
version = "0.1.0"
description = "Unified PET simulation pipeline across MCGPU-PET and GATE"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["petsim"]
